Fix hang on first ssh connect
Use Open3.capture2 instead of IO.popen. When ssh is configured to create
a new master connection on first connect, the ssh process will fork a
new one, and continue to run the command in that one. Therefore our
previous code was waiting indefinitely for the ssh command to finish.